NO se puede delete cancion si se está reproduciendo (reunión)

+++ src/presentation/controller/DeleteMusicController.java
@@ -1,6 +1,7 @@
 package presentation.controller;
 
 import business.SongManager;
+import presentation.view.BottomBarPanel;
 import presentation.view.DeleteMusicPanelView;
 
 import javax.swing.*;
@@ -17,7 +18,7 @@ public class DeleteMusicController implements ActionListener {
     @Override
     public void actionPerformed(ActionEvent e) {
         if (e.getActionCommand().equals(DeleteMusicPanelView.DELETE)) {
-            if (deleteMusicPanelView.getNameSong() != null) {
+            if (deleteMusicPanelView.getNameSong() != null && !BottomBarPanel.getSong().getTitle().equals(deleteMusicPanelView.getNameSong())) {
                 SongManager.DeleteSong(deleteMusicPanelView.getNameSong());
                 JOptionPane.showMessageDialog(deleteMusicPanelView, "DELETE SUCCESSFULLY", "OKAY", JOptionPane.INFORMATION_MESSAGE);
             } else {
